pyproject.toml
[tool.poetry]
name = "speaker-verification"
version = "0.1.0"
description = ""
authors = ["Daniel Neururer <neud@zhaw.ch>"]
readme = "README.md"
packages = [{include = "speaker_verification"}]
repository = "https://github.com/zhaw-nerd/SpeakerVerification"

[tool.poetry.dependencies]
python = "^3.10,<3.12"
torch = ">=1.7.0"
torchaudio = ">=0.7.0"
asteroid_filterbanks = "==0.4.0"
numpy = "^1.24.1"
soundfile = "^0.11.0"
scipy = "^1.10.0"
scikit-learn = "^1.2.0"
pyyaml = "^6.0"
h5py = "^3.7.0"
requests = "^2.28.2"
pydantic = {extras = ["dotenv"], version = "^1.10.4"}
psutil = "^5.9.4"
rich = "^13.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
